moving ListenerConfig to separate file
